Moved and generalized IsolatedNetworkManager to common repository
