+++ vantage6/node/__init__.py
@@ -30,8 +30,8 @@ from socketio import ClientNamespace, Client as SocketIO
 from gevent.pywsgi import WSGIServer
 from enum import Enum
 
-from vantage6.common.docker_addons import (
-    ContainerKillListener, check_docker_running
+from vantage6.common.docker.addons import (
+    ContainerKillListener, check_docker_running, running_in_docker
 )
 from vantage6.common.globals import VPN_CONFIG_FILE
 from vantage6.cli.context import NodeContext
@@ -41,7 +41,7 @@ from vantage6.node.server_io import NodeClient
 from vantage6.node.proxy_server import app
 from vantage6.node.util import logger_name
 from vantage6.node.docker.docker_manager import DockerManager
-from vantage6.node.docker.network_manager import NetworkManager
+from vantage6.common.docker.network_manager import NetworkManager
 from vantage6.node.docker.vpn_manager import VPNManager
 
 
@@ -182,8 +182,14 @@ class Node(object):
         self.connect_to_socket()
 
         # setup docker isolated network manager
-        isolated_network_mgr = \
-            NetworkManager(self.ctx.docker_network_name)
+        internal_ = running_in_docker()
+        if not internal_:
+            self.log.warn(
+                "Algorithms have internet connection! "
+                "This happens because you use 'vnode-local'!"
+            )
+        isolated_network_mgr = NetworkManager(self.ctx.docker_network_name)
+        isolated_network_mgr.create_network(is_internal=internal_)
 
         # Setup tasks dir
         self._set_task_dir(self.ctx)
