+++ vantage6/node/docker/docker_manager.py
@@ -16,11 +16,11 @@ from typing import Dict, List, NamedTuple, Union
 from pathlib import Path
 
 from vantage6.common.globals import APPNAME
+from vantage6.common.docker.addons import running_in_docker
 from vantage6.node.docker.docker_base import DockerBaseManager
-from vantage6.node.docker.utils import running_in_docker
 from vantage6.node.docker.vpn_manager import VPNManager
 from vantage6.node.util import logger_name
-from vantage6.node.docker.network_manager import NetworkManager
+from vantage6.common.docker.network_manager import NetworkManager
 from vantage6.node.docker.task_manager import DockerTaskManager
 
 log = logging.getLogger(logger_name(__name__))
