+++ vantage6/node/docker/task_manager.py
@@ -7,15 +7,14 @@ from typing import Dict, List
 from pathlib import Path
 
 from vantage6.common.globals import APPNAME
-from vantage6.common.docker_addons import (
+from vantage6.common.docker.addons import (
     remove_container_if_exists, remove_container
 )
 from vantage6.node.util import logger_name
 from vantage6.node.docker.vpn_manager import VPNManager
-from vantage6.node.docker.network_manager import NetworkManager
+from vantage6.common.docker.network_manager import NetworkManager
 from vantage6.node.docker.docker_base import DockerBaseManager
-from vantage6.node.docker.utils import running_in_docker
-from vantage6.common.docker_addons import pull_if_newer
+from vantage6.common.docker.addons import pull_if_newer, running_in_docker
 
 
 class DockerTaskManager(DockerBaseManager):
